更新安卓sdk代码

+++ android/src/main/kotlin/com/example/pp_bluetooth_kit_flutter/extension/PPLefuBleExtensionData.kt
@@ -215,7 +215,7 @@ fun PPLefuBleConnectManager.sendCommonState(state: Boolean, callBack: Result?) {
     callBack?.success(dict)
 }
 
-fun PPLefuBleConnectManager.sendDfuResult(progress: Float, isSuccess: Boolean) {
+fun PPLefuBleConnectManager.sendDfuResult(progress: Double, isSuccess: Boolean) {
     val dict = mapOf(
         "progress" to progress,
         "isSuccess" to isSuccess
